create libs/test-decks and move createPrecinctSummaryBallotTestDeck
We need to factor this function out because it's needed in VxMark and VxDesign.
Because it imports libs/printing, which in turn imports libs/utils,
we can't put this in libs/utils like other test deck-related utils
without creating a circular dependency.

For consistency, future commits will move other shared test deck utils
into libs/test-decks.
