+++ libs/test-decks/.eslintignore
@@ -1,0 +1,3 @@
+build
+coverage
+vitest.config.ts
