+++ libs/test-decks/.lintstagedrc.js
@@ -1,0 +1,5 @@
+const { base } = require('../../.lintstagedrc.shared');
+
+module.exports = {
+  ...base,
+};
