+++ libs/test-decks/package.json
@@ -1,0 +1,53 @@
+{
+  "name": "@votingworks/test-decks",
+  "version": "1.0.0",
+  "private": true,
+  "description": "Rendering of summary (BMD) ballot test decks",
+  "license": "GPL-3.0-only",
+  "main": "build/index.js",
+  "types": "build/index.d.js",
+  "scripts": {
+    "build": "is-ci build:ci build:dev",
+    "build:ci": "pnpm --filter $npm_package_name... --sequential build:self",
+    "build:dev": "pnpm --filter $npm_package_name... build:self",
+    "build:self": "tsgo --build tsconfig.build.json",
+    "clean": "pnpm --filter $npm_package_name... clean:self",
+    "clean:self": "rm -rf build && tsgo --build --clean tsconfig.build.json",
+    "lint": "pnpm type-check && eslint --cache .",
+    "lint:fix": "pnpm type-check && eslint --cache . --fix",
+    "test": "is-ci test:ci test:watch",
+    "test:ci": "pnpm build && TZ=America/Anchorage pnpm vitest run --coverage",
+    "test:coverage": "TZ=America/Anchorage vitest --coverage",
+    "test:run": "TZ=America/Anchorage vitest run",
+    "test:watch": "TZ=America/Anchorage vitest",
+    "type-check": "tsgo --build"
+  },
+  "dependencies": {
+    "@votingworks/basics": "workspace:*",
+    "@votingworks/hmpb": "workspace:*",
+    "@votingworks/printing": "workspace:*",
+    "@votingworks/types": "workspace:*",
+    "@votingworks/ui": "workspace:*",
+    "@votingworks/utils": "workspace:*",
+    "react": "18.3.1",
+    "react-dom": "18.3.1",
+    "styled-components": "^5.3.11"
+  },
+  "devDependencies": {
+    "@types/node": "20.17.31",
+    "@types/react": "18.3.3",
+    "@types/react-dom": "18.3.0",
+    "@types/styled-components": "^5.1.26",
+    "@vitest/coverage-istanbul": "^4.1.6",
+    "@votingworks/fixtures": "workspace:*",
+    "@votingworks/image-utils": "workspace:*",
+    "esbuild": "0.21.2",
+    "esbuild-runner": "2.2.2",
+    "eslint-plugin-vx": "workspace:*",
+    "is-ci-cli": "2.2.0",
+    "jest-image-snapshot": "^6.4.0",
+    "sort-package-json": "^1.50.0",
+    "vitest": "^4.1.6"
+  },
+  "packageManager": "pnpm@9.15.9"
+}
