+++ libs/test-decks/src/index.ts
@@ -1,0 +1,1 @@
+export * from './summary_ballots';
