+++ libs/test-decks/test/setupTests.ts
@@ -1,0 +1,27 @@
+import { afterAll, beforeAll, expect } from 'vitest';
+import {
+  clearTemporaryRootDir,
+  setupTemporaryRootDir,
+} from '@votingworks/fixtures';
+import {
+  ToMatchPdfSnapshotOptions,
+  buildToMatchPdfSnapshot,
+} from '@votingworks/image-utils';
+import { toMatchImageSnapshot } from 'jest-image-snapshot';
+
+declare global {
+  // eslint-disable-next-line @typescript-eslint/no-namespace
+  namespace jest {
+    interface Matchers<R> {
+      toMatchPdfSnapshot(options?: ToMatchPdfSnapshotOptions): Promise<R>;
+    }
+  }
+}
+
+expect.extend({
+  toMatchImageSnapshot,
+  toMatchPdfSnapshot: buildToMatchPdfSnapshot(expect),
+});
+
+beforeAll(setupTemporaryRootDir);
+afterAll(clearTemporaryRootDir);
